+++ src/TaskRunnerExtended/ToolWindows/TaskRunnerToolWindow.cs
@@ -32,15 +32,15 @@ public class TaskRunnerToolWindow : ToolWindow
     {
         Children =
         [
-            ToolbarChild.Command<ShowTasksCommand>(),
-            ToolbarChild.Command<ShowBackgroundCommand>(),
-            ToolbarChild.Command<ShowFeedbackCommand>(),
-            ToolbarChild.Separator,
             ToolbarChild.Command<RefreshCommand>(),
             ToolbarChild.Separator,
             ToolbarChild.Command<StopAllCommand>(),
             ToolbarChild.Separator,
             ToolbarChild.Command<CollapseAllCommand>(),
+            ToolbarChild.Separator,
+            ToolbarChild.Command<ShowTasksCommand>(),
+            ToolbarChild.Command<ShowBackgroundCommand>(),
+            ToolbarChild.Command<ShowFeedbackCommand>(),
         ],
     };